--- src/components/landingContent/content.tsx
@@ -1,18 +1,0 @@
-// Adoption Directions
-export const ADOPTION_DIRECTIONS_HEADER = {
-  ENG: 'How to Adopt a Tree',
-};
-
-export const ADOPTION_DIRECTIONS = {
-  ENG: [
-    'Click “Sign-Up” at the top right corner to create an account.',
-    'Once you’re logged in, enter your address in the search bar on the map to find adoptable trees near you.',
-    'Click on the icon of the tree you’d like to adopt, either the blue triangle for a newly planted tree ' +
-      'or a green circle for an older tree.',
-    'Click on “More Info”.',
-    'Click on the “Adopt” button on the next screen to adopt that tree. It will be added to the “My Trees” ' +
-      'list so you can easily find your tree and record your tree care activities.',
-    'Record each activity by clicking on “My Trees”, finding the tree you adopted, and submitting the ' +
-      'tree care activity on that tree’s page. It’s easy!',
-  ],
-};
