+++ src/components/landingContent/index.tsx
@@ -1,10 +1,11 @@
 import OrderedList from '../orderedList';
 import { Typography } from 'antd';
-import { ADOPTION_DIRECTIONS, ADOPTION_DIRECTIONS_HEADER } from './content';
 import React from 'react';
 import {
   QUESTIONS_DIRECTIONS,
   LANDING_BODY,
+  ADOPTION_DIRECTIONS,
+  ADOPTION_DIRECTIONS_HEADER,
 } from '../../containers/landing/content';
 import { Languages } from '../../App';
