+++ src/containers/landing/content.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
+import { CONTACT_EMAIL } from '../../assets/links';
 
 // Landing
 export const LANDING_TITLE = {
   ENG: "Boston's Street Trees",
 };
-
 export const QUESTIONS_DIRECTIONS = {
   ENG: [
     <>
@@ -22,7 +22,7 @@ export const QUESTIONS_DIRECTIONS = {
     <>
       If not, reach out to us at{' '}
       <a
-        href={'mailto:adopt@treeboston.org'}
+        href={`mailto:${CONTACT_EMAIL}`}
         target={'_blank'}
         rel={'noopener noreferrer'}
       >
@@ -32,7 +32,6 @@ export const QUESTIONS_DIRECTIONS = {
     </>,
   ],
 };
-
 export const LANDING_BODY = {
   ENG: (
     <>
@@ -75,4 +74,22 @@ export const MODAL_PARAGRAPH = {
 };
 export const MODAL_OK_TEXT = {
   ENG: "Explore Boston's urban forest!",
+};
+
+// Adoption Directions
+export const ADOPTION_DIRECTIONS_HEADER = {
+  ENG: 'How to Adopt a Tree',
+};
+export const ADOPTION_DIRECTIONS = {
+  ENG: [
+    'Click “Sign-Up” at the top right corner to create an account.',
+    'Once you’re logged in, enter your address in the search bar on the map to find adoptable trees near you.',
+    'Click on the icon of the tree you’d like to adopt, either the blue triangle for a newly planted tree ' +
+      'or a green circle for an older tree.',
+    'Click on “More Info”.',
+    'Click on the “Adopt” button on the next screen to adopt that tree. It will be added to the “My Trees” ' +
+      'list so you can easily find your tree and record your tree care activities.',
+    'Record each activity by clicking on “My Trees”, finding the tree you adopted, and submitting the ' +
+      'tree care activity on that tree’s page. It’s easy!',
+  ],
 };
